refactor: use 32-bit mode arguments

+++ src/reader/internal/metadata_types.cpp
@@ -185,7 +185,7 @@ unpack_directories(logger& lgr, global_metadata::Meta const& meta) {
 }
 
 // TODO: merge with inode_rank in metadata_v2
-int mode_rank(uint16_t mode) {
+int mode_rank(uint32_t mode) {
   switch (posix_file_type::from_mode(mode)) {
   case posix_file_type::directory:
     return 0;
